Created class BitsTestCase which inherits from unittest.TestCase
