+++ test/helper.py
@@ -3,9 +3,12 @@
 (Didn't want to also call it util.py)
 """
 import time
-
+import unittest
 
 # https://medium.com/pythonhive/python-decorator-to-measure-the-execution-time-of-methods-fa04cb6bb36d
+import bitstring
+
+
 def timeit(method):
     """Decorator to print execution time of method."""
 
@@ -21,3 +24,17 @@ def timeit(method):
         return result
 
     return timed
+
+
+class BitsTestCase(unittest.TestCase):
+    """Custom test case class to reuse testing patterns."""
+
+    def setUp(self):
+        super(BitsTestCase, self).setUp()
+
+        def assert_bit(b: bitstring.Bits, value: int, length: int):
+            """Custom bit assertion function."""
+            self.assertEqual(b.uint, value)
+            self.assertEqual(len(b), length)
+
+        self.assertBit = assert_bit
