+++ test/util/test_bitseq_from_str.py
@@ -1,21 +1,12 @@
-import unittest
-
 import bitstring
 
 # noinspection PyUnresolvedReferences
 import test.context
+from test.helper import BitsTestCase
 from util.bitseq import bitseq_from_str
 
 
-class TestBitSeq(unittest.TestCase):
-
-    def setUp(self):
-        def assert_bit(b: bitstring.Bits, value: int, length: int):
-            """Custom bit assertion function."""
-            self.assertEqual(b.uint, value)
-            self.assertEqual(len(b), length)
-
-        self.assertBit = assert_bit
+class TestBitSeq(BitsTestCase):
 
     def test_bitseq_from_str_returns_bitstring_Bits_instance(self):
         b = bitseq_from_str("0x1", "0x2", "0x3", "0x4")
